backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15
django-cors-headers>=4.3
gunicorn>=22.0
weasyprint>=62.0
